[refactor][plugin] refactor the class package of keyguard hooks.
