+++ app/src/main/java/com/tianma/tweaks/miui/xp/hook/systemui/keyguard/v20190507/ChooseKeyguardClockActivityHook.kt
@@ -1,4 +1,4 @@
-package com.tianma.tweaks.miui.xp.hook.systemui.keyguard
+package com.tianma.tweaks.miui.xp.hook.systemui.keyguard.v20190507
 
 import android.content.Context
 import android.content.Intent
