+++ app/src/main/java/com/tianma/tweaks/miui/xp/hook/systemui/keyguard/v20190507/MiuiKeyguardBaseClockHook.java
@@ -1,4 +1,4 @@
-package com.tianma.tweaks.miui.xp.hook.systemui.keyguard;
+package com.tianma.tweaks.miui.xp.hook.systemui.keyguard.v20190507;
 
 import static com.tianma.tweaks.miui.xp.wrapper.XposedWrapper.hookAllConstructors;
