+++ app/src/main/java/com/tianma/tweaks/miui/xp/hook/systemui/keyguard/v20191213/MiuiLeftTopClockHook.kt
@@ -1,4 +1,4 @@
-package com.tianma.tweaks.miui.xp.hook.systemui.keyguard
+package com.tianma.tweaks.miui.xp.hook.systemui.keyguard.v20191213
 
 import android.view.View
 import android.view.ViewTreeObserver.OnWindowAttachListener
